requirements.txt
django>=4.2.0
google-generativeai>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
django-crispy-forms>=2.0
whitenoise>=6.0.0
PyMuPDF==1.22.5
tiktoken>=0.9.0
requests>=2.31.0
PyPDF2>=3.0.0
